optimize message center websocket
